refactor(media): move orphan selection logic into shared core
scanUploadDir wandert mit, obwohl es node:fs anfasst: Es trägt den Ausschluss
von _old_uploads, das Überspringen von Punktdateien und die ENOENT-Toleranz —
eine Zweitfassung im Endpunkt hätte den Altbestand der Migration gelöscht.
