+++ tsconfig.json
@@ -10,6 +10,7 @@
 		"sourceMap": true,
 		"strict": true,
 		"moduleResolution": "bundler",
+		"allowImportingTsExtensions": true,
 		"noImplicitReturns": true,
 		"noImplicitOverride": true,
 		"noUncheckedIndexedAccess": true,
